修复ProtoBufSerializer在Magi模式下不兼容问题

+++ src/main/java/com/jarvis/cache/serializer/protobuf/WriteByteBuf.java
@@ -44,7 +44,7 @@ public class WriteByteBuf {
     }
 
 
-    public void writeByte(int value) {
+    public void writeByte(byte value) {
         int length = 1;
         hasLength(length);
         HeapByteBufUtil.setByte(array, writerIndex, value);
